Move auth.ts
