+++ src/lib/auth.ts
@@ -1,7 +1,7 @@
 import NextAuth from 'next-auth';
 import Google from 'next-auth/providers/google';
 import { MongoDBAdapter } from '@auth/mongodb-adapter';
-import client from './lib/mongodb';
+import client from './mongodb';
 
 export const { auth, handlers, signIn, signOut } = NextAuth({
     adapter: MongoDBAdapter(client),
